requirements.txt
httpx==0.27.2
PyJWT[crypto]==2.9.0
python-dateutil==2.9.0.post0
